DOC: Add examples to the SQL docs (#31633)
Significantly improve the example snippets in the documentation.
The examples are part of the test suite and checked nightly.
To help readability, the existing dataset was extended (test_emp renamed
to emp plus library).
Improve output of JDBC tests to be consistent with the CLI
Add lenient flag to JDBC asserts to allow type widening (a long is
equivalent to a integer as long as the value is the same).

(cherry picked from commit de9e56aa0161abb30047e7d96b814ad250b69165)

+++ x-pack/plugin/sql/jdbc/src/main/java/org/elasticsearch/xpack/sql/jdbc/jdbc/TypeConverter.java
@@ -61,9 +61,9 @@ final class TypeConverter {
     
     static {
         Map<Class<?>, JDBCType> aMap = Arrays.stream(DataType.values())
-                .filter(dataType -> dataType.javaClass() != null 
-                        && dataType != DataType.HALF_FLOAT 
-                        && dataType != DataType.SCALED_FLOAT 
+                .filter(dataType -> dataType.javaClass() != null
+                        && dataType != DataType.HALF_FLOAT
+                        && dataType != DataType.SCALED_FLOAT
                         && dataType != DataType.TEXT)
                 .collect(Collectors.toMap(dataType -> dataType.javaClass(), dataType -> dataType.jdbcType));
         // apart from the mappings in {@code DataType} three more Java classes can be mapped to a {@code JDBCType.TIMESTAMP}
@@ -428,7 +428,7 @@ final class TypeConverter {
             case SMALLINT:
             case INTEGER:
             case BIGINT:
-                return Float.valueOf((float) ((Number) val).longValue());
+                return Float.valueOf(((Number) val).longValue());
             case REAL:
             case FLOAT:
             case DOUBLE:
@@ -447,7 +447,7 @@ final class TypeConverter {
             case SMALLINT:
             case INTEGER:
             case BIGINT:
-                return Double.valueOf((double) ((Number) val).longValue());
+                return Double.valueOf(((Number) val).longValue());
             case REAL:
             case FLOAT:
             case DOUBLE:
